renamed Wizard to BaseWizard
